backend/pyproject.toml
[tool.poetry]
name = "sbc-backend"
version = "0.1.0"
description = ""
authors = ["Lawrence Ephrim <ephrimlawrence0@gmail.com>"]
readme = "README.md"
packages = [{ include = "sbc_backend" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.101.1"
boto3 = "^1.28.30"
mangum = "^0.17.0"
openai = "^0.27.8"
pg8000 = "^1.30.1"
python-jose = "^3.3.0"
python-dotenv = "^1.0.0"
sentry-sdk = "^1.29.2"
sqlalchemy = "^2.0.20"
sqlalchemy-easy-softdelete = "^0.8.2"
pydantic-settings = "^2.0.3"
psycopg2-binary = "^2.9.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
